+++ src/pydocstyle/utils.py
@@ -6,7 +6,7 @@ from itertools import tee, zip_longest
 from typing import Any, Iterable, Tuple
 
 # Do not update the version manually - it is managed by `bumpversion`.
-__version__ = '6.0.1rc'
+__version__ = '6.1.2rc'
 log = logging.getLogger(__name__)
 
 #: Regular expression for stripping non-alphanumeric characters